refactor of ted state tracker
