+++ models/datamodule/TedDataModule.py
@@ -1,4 +1,3 @@
-import copy
 import os
 import pickle
 from abc import ABC
@@ -7,7 +6,7 @@ import pandas as pd
 import torch
 from typing import Optional, List
 import numpy as np
-import pytorch_lightning as pl
+from sklearn.preprocessing import LabelEncoder
 from torch.utils.data import DataLoader
 
 from models.datamodule.DataModule import DataModule
@@ -61,39 +60,56 @@ class TedDataModule(DataModule, ABC):
     def __init__(
             self,
             df: pd.DataFrame,
-            train: np.array,
-            label_train: np.array,
-            set_labels_train: np.array,
-            val: np.array,
-            label_val: np.array,
-            set_labels_val: np.array,
-            test: np.array,
-            label_test: np.array,
-            set_labels_test: np.array,
-            train_indexes,
-            validation_indexes,
-            test_indexes,
             batch_size: int = 32
     ):
         super().__init__(df)
+        train, train_labels, set_train_labels, train_indexes = self._get_samples_by_dataset(df, 'train')
+        validation, validation_labels, set_validation_labels, validation_indexes = self._get_samples_by_dataset(df, 'dev')
+        test, test_labels, set_test_labels, test_indexes = self._get_samples_by_dataset(df, 'test')
         self.sgd_test = None
         self.sgd_val = None
         self.sgd_train = None
         self.batch_size = batch_size
         self.train = train
-        self.label_train = label_train
-        self.set_labels_train = set_labels_train
-        self.val = val
-        self.label_val = label_val
-        self.set_labels_val = set_labels_val
+        self.label_train = train_labels
+        self.set_labels_train = set_train_labels
+        self.val = validation
+        self.label_val = validation_labels
+        self.set_labels_val = set_validation_labels
         self.test = test
-        self.label_test = label_test
-        self.set_labels_test = set_labels_test
+        self.label_test = test_labels
+        self.set_labels_test = set_test_labels
         self.train_indexes = train_indexes
         self.validation_indexes = validation_indexes
         self.test_indexes = test_indexes
         self._num_features = train.shape[1]
-        self._num_classes = len(np.unique(label_train))
+        self._num_classes = len(np.unique(train_labels))
+
+    def _get_samples_by_dataset(self, df_state: pd.DataFrame, type_data: str) -> tuple:
+        action_encoder = LabelEncoder()
+        action_encoder.fit(df_state['Label'])
+        dataset = df_state[df_state['Type'] == type_data]
+        indexes = dataset['Index'].to_list()
+        labels = action_encoder.transform(dataset['Label'])
+        set_labels = [
+            list(set(action_encoder.transform(actions)))
+            for actions in dataset['Set_Label']
+        ]
+        max_len_of_set_labels = max([len(set_of_label) for set_of_label in set_labels])
+        set_labels = [
+            set_of_label + [-1] * (max_len_of_set_labels - len(set_of_label))
+            for set_of_label in set_labels
+        ]
+        features = np.array(dataset['State'].to_list())
+
+        assert len(indexes) == len(features) == len(labels) == len(set_labels), \
+            f"Bab dimension of features: {len(features)} - " \
+            f"labels : {len(labels)} - set_labels: {len(set_labels)} " \
+            f"index: {len(indexes)}"
+        numbers = [len(set_label) for set_label in set_labels]
+        assert max(numbers) == min(numbers), "Bab dimension of set_labels"
+
+        return features, labels, set_labels, indexes
 
     def setup(self, stage: Optional[str] = None):
